refactor(compiler): move authenticator attribute to dedicated file
